Adjustments ready for tomorrow
I got rid of everything we didn't need from Watson because I found out the white sox and red sox wouldn't show up for some reason. After getting rid of all the cities in Watson I hardcoded the white and red sox to work. I think it's an issue with spaces in the team name. Not sure where else that will come up but we can go over it tomorrow